refactor(types): merge action arg types to same scope
