+++ packages/graz/src/index.ts
@@ -14,4 +14,3 @@ export * from "./provider";
 export * from "./provider/events";
 export * from "./types/core";
 export * from "./types/keplr";
-export * from "./types/methods";
